Many improvements in settings and strings

+++ applications_user/wmbus_decoder/protocol/decode/wmbus_decode.h
@@ -7,7 +7,7 @@
 bool wmbus_decode_3of6_bits(
     const uint8_t* raw,
     size_t raw_bit_len,
-    uint8_t bit_offset,
+    size_t bit_offset,
     uint8_t* out,
     size_t out_max,
     size_t* out_len);
